Clean up per customer activity page.

+++ db/sessions_per_customer_stats.sql
@@ -3,8 +3,8 @@ SELECT
     count(*)::integer as config_request_count,
     count(used_turn OR NULL)::integer as used_turn_count,
     count(activated OR NULL)::integer as session_connected_count,
-    sum(CASE WHEN type = 'desktop' and activated = true THEN 1 ELSE 0 END) as web_count,
-    sum(CASE WHEN type = 'mobile' and activated = true THEN 1 ELSE 0 END) as mobile_count,
+    sum(CASE WHEN type = 'desktop' and activated = true THEN 1 ELSE 0 END)::integer as web_count,
+    sum(CASE WHEN type = 'mobile' and activated = true THEN 1 ELSE 0 END)::integer as mobile_count,
     max(created_at) as last_activity,
     CASE
       WHEN ended_at IS NULL THEN true
